refactor: codebase 2.0
added UnitTest project
