+++ GettingReal/Application/ViewModels/UILogicClasses.cs
