+++ GettingReal/Domain/Models/Audio/Lecture.cs
