+++ GettingReal/Domain/Models/Metadata/AudioMetadata.cs
