+++ GettingReal/GettingReal.sln
@@ -3,11 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34511.84
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "1.UI", "GettingReal.UI\1.UI.csproj", "{D986DD5F-17E6-4ECB-9CF9-5E418F0D8335}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Application", "Application\Application.csproj", "{57C380D5-2A2A-49DF-A276-21EA144A84B6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "2.Application", "Application\2.Application.csproj", "{57C380D5-2A2A-49DF-A276-21EA144A84B6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain", "Domain\Domain.csproj", "{CF8EFF83-E90B-4468-9519-9E8C165D6884}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "3.Domain", "Domain\3.Domain.csproj", "{CF8EFF83-E90B-4468-9519-9E8C165D6884}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UI", "UI\UI.csproj", "{A63B2C6B-78A4-4C4A-B18C-5F452D619868}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTest", "UnitTest\UnitTest.csproj", "{8A2EDDC7-1E9F-46E3-8ADE-250226B2E00D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,10 +17,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D986DD5F-17E6-4ECB-9CF9-5E418F0D8335}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D986DD5F-17E6-4ECB-9CF9-5E418F0D8335}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D986DD5F-17E6-4ECB-9CF9-5E418F0D8335}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D986DD5F-17E6-4ECB-9CF9-5E418F0D8335}.Release|Any CPU.Build.0 = Release|Any CPU
 		{57C380D5-2A2A-49DF-A276-21EA144A84B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{57C380D5-2A2A-49DF-A276-21EA144A84B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{57C380D5-2A2A-49DF-A276-21EA144A84B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -27,6 +25,14 @@ Global
 		{CF8EFF83-E90B-4468-9519-9E8C165D6884}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CF8EFF83-E90B-4468-9519-9E8C165D6884}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CF8EFF83-E90B-4468-9519-9E8C165D6884}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A63B2C6B-78A4-4C4A-B18C-5F452D619868}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A63B2C6B-78A4-4C4A-B18C-5F452D619868}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A63B2C6B-78A4-4C4A-B18C-5F452D619868}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A63B2C6B-78A4-4C4A-B18C-5F452D619868}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8A2EDDC7-1E9F-46E3-8ADE-250226B2E00D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8A2EDDC7-1E9F-46E3-8ADE-250226B2E00D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8A2EDDC7-1E9F-46E3-8ADE-250226B2E00D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8A2EDDC7-1E9F-46E3-8ADE-250226B2E00D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
