+++ GettingReal/UI/App.xaml
@@ -1,8 +1,8 @@
-﻿<Application x:Class="GettingReal.UI.App"
+﻿<Application x:Class="UI.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:GettingReal.UI"
-             StartupUri="Views\MainWindow.xaml">
+             xmlns:local="clr-namespace:UI"
+             StartupUri="Views/MainWindow.xaml">
     <Application.Resources>
          
     </Application.Resources>
