+++ GettingReal/UI/AssemblyInfo.cs
