+++ GettingReal/UI/UI.csproj
@@ -8,8 +8,4 @@
     <UseWPF>true</UseWPF>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Application\2.Application.csproj" />
-  </ItemGroup>
-
 </Project>
